requirements.txt
django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pre-commit==3.3.2
geopy==2.3.0
django-cron==0.6.0
postgis==1.0.4
